scripts/requirements.txt
google-genai>=0.5.0
Pillow>=10.0.0
